% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Createh2D2Object.R
\name{Createh2D2Object}
\alias{Createh2D2Object}
\title{Create an h2D2 object.}
\usage{
Createh2D2Object(
  betaHat,
  sigmaHat,
  R,
  SNP_ID = NULL,
  trait = "quantitative",
  N = NULL,
  N1 = NULL,
  N0 = NULL,
  a = 0.005,
  b = 2e+05,
  coverage = 0.95,
  purity = 0.5
)
}
\arguments{
\item{betaHat}{M-vector of standardized effect sizes. Per-allele effect sizes
should be multiplied by sqrt(2\*MAF\*(1-MAF)) before input.}

\item{sigmaHat}{M-vector of standard errors of standardized effect sizes.
Standard errors of per-allele effect sizes should be multiplied by 
sqrt(2\*MAF\*(1-MAF)) before input.}

\item{R}{An M-by-M LD matrix that can be coerced to "dsCMatrix".
The diagonal elements should be set as 0s.}

\item{SNP_ID}{Identifiers of SNPs. The default is c("SNP_1", ...).}

\item{trait}{One of "quantitative" or "binary".}

\item{N}{The sample size. Required for quantitative traits.}

\item{N1}{Number of cases. Required for binary traits.}

\item{N0}{Number of controls. Required for binary traits.}

\item{a}{Shape parameters for sigma^2. Either a positive real number or an
M-vector of positive real numbers.}

\item{b}{Shape parameters for 1-h^2. A positive real number.}

\item{coverage}{A number between 0 and 1 specifying the "coverage"
of the credible sets.}

\item{purity}{A number between 0 and 1 specifying the minimum 
absolute correlation allowed in a credible set.}
}
\description{
Create an h2D2 object from GWAS summary statistics and an LD matrix estimate.
GWAS summary data, h2-D2 hyper-parameters, and useful variables are stored.
}
