% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credible_sets.R
\name{credible_sets}
\alias{credible_sets}
\title{Identify credible sets for h2D2 object.}
\usage{
credible_sets(h2D2, coverage = 0.95, purity = 0.5)
}
\arguments{
\item{h2D2}{An h2D2 object with MCMC samples.}

\item{coverage}{A number between 0 and 1 specifying the "coverage"
of the credible sets.}

\item{purity}{A number between 0 and 1 specifying the minimum 
absolute correlation allowed in a credible set.}
}
\description{
A greedy search algorithm to identify all credible sets achieving target
"coverage" with minimum absolute correlation not less than "purity".
}
