% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2D2_CS.R
\name{h2D2_CS}
\alias{h2D2_CS}
\title{Identify credible sets for h2D2 object.}
\usage{
h2D2_CS(h2D2, coverage = 0.95, purity = 0.5)
}
\arguments{
\item{h2D2}{An h2D2 object with MCMC samples.}

\item{coverage}{A number between 0 and 1 specifying the required coverage
of the credible sets.}

\item{purity}{A number between 0 and 1 specifying the minimum 
absolute correlation allowed in a credible set.}
}
\value{
A list with the following elements:
\describe{
  \item{sets}{A list in which each element is a vector containing the
  indices of the variables in the CS.}
  \item{CL}{The credible level of each CS.}
  \item{purity}{The purity of each CS.}
}
}
\description{
A greedy search algorithm to identify all credible sets achieving target
"coverage" with minimum absolute correlation not less than "purity".
}
