% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2D2_MCMC.R
\name{h2D2_MCMC}
\alias{h2D2_MCMC}
\title{MCMC sampling under h2-D2 prior.}
\usage{
h2D2 = h2D2_MCMC(h2D2, mcmc_n = 10000, burn_in = 5000, 
                 thin = 1, stepsize = 2, seed = 428)
}
\arguments{
\item{h2D2}{An h2D2 object.}

\item{mcmc_n}{Number of posterior samples to generate.}

\item{burn_in}{Number of early samples to discard.}

\item{thin}{Thinning parameter of the MCMC chain. The default is 1, 
indicating no thinning.}

\item{stepsize}{Step size for proposal.}

\item{seed}{Random seed for MCMC sampling.}

\item{pre_mcmc_n}{If b = NULL, a pre-training process will be used to
estimate b before MCMC. This parameter specifies the number of MCMC
iterations of each training.}

\item{pre_use}{The number of samples used to estimate b in each pre 
training.}

\item{pre_p}{p value threshold to determine whether the pre training can
be stopped.}

\item{pre_maxiter}{Maximum number of pre-training iterations.}
}
\value{
An h2D2 object with MCMC samples. See \code{\link{h2D2-class}}.
}
\description{
Using MCMC sampler to fit posterior distribution under h2-D2 prior.
}
