% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDL.R
\name{HDL}
\alias{HDL}
\title{Estimate local heritability using high-definition likelihood method.}
\usage{
HDL(z, N, R_eig = R_eig)
HDL(z, N, R = R)
}
\arguments{
\item{z}{M-vector of z-scores.}

\item{N}{GWAS sample size. For binary traits, 'N' is the sum of the
number of cases and the number of controls.}

\item{R}{An M-by-M LD matrix with all diagonal elements zeros.
The diagonal elements will be coerced to zeros.}

\item{R_eig}{Spectral decomposition of LD matrix. A list return by 
\code{\link[eigen]{base}} with components 'values' and 'vectors'. The
diagonal elements of LD matrix are zeros.}

\item{tol}{A numeric tolerance. Eigenvalues smaller than this value will be
removed.}
}
\value{
A numeric value of local heritability estimate.
}
\description{
Obtain maximum likelihood estimate of local heritability using z-scores and
LD matrix.
}
\references{
https://www.nature.com/articles/s41588-020-0653-y
}
