% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rinvGauss}
\alias{rinvGauss}
\title{The inverse Gaussian distribution}
\usage{
rinvGauss(mu, lambda)
}
\arguments{
\item{mu}{The mean parameter. Must be positive.}

\item{lambda}{The shape parameter. Must be positive.}
}
\value{
A random number from inverse Gaussian distribution.
}
\description{
Random generation for the inverse Gaussian distribution with 
parameters \\code{mu} and \\code{lambda}.
}
\details{
The inverse Gaussian distribution with parameters $mean = \\mu$ and 
$shape = \\lambda$ has density
$$
f(x) = \\sqrt\{\\frac\{\\lambda\}\{2 \\pi x^3\}\} 
\\exp \{ -\\frac\{\\lambda (x-\\mu)^2\}\{2 \\mu^2 x\} \}
$$
}
