% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Createh2D2Object.R
\name{Createh2D2Object}
\alias{Createh2D2Object}
\title{Create an h2D2 object.}
\usage{
For quantitative traits,
h2D2 = Createh2D2Object(z,
                        R,
                        N,
                        SNP_ID = NULL,
                        trait = "quantitative",
                        in_sample_LD = F,
                        a = 0.005,
                        coverage = 0.95,
                        purity = 0.5)

For binary traits,
h2D2 = Createh2D2Object(z,
                        R,
                        N1,
                        N0,
                        SNP_ID = NULL,
                        trait = "binary",
                        in_sample_LD = F,
                        a = 0.005,
                        coverage = 0.95,
                        purity = 0.5)
}
\arguments{
\item{z}{M-vector of z-scores.}

\item{R}{An M-by-M LD matrix that can be coerced to 
\code{\link[Matrix]{dsCMatrix-class}}.
The diagonal elements should be set as 0s.}

\item{N}{The sample size. Required for quantitative traits.}

\item{N1}{Number of cases. Required for binary traits.}

\item{N0}{Number of controls. Required for binary traits.}

\item{SNP_ID}{Identifiers of SNPs. The default is c("SNP_1", ...).}

\item{trait}{One of "quantitative" or "binary".}

\item{in_sample_LD}{Whether the LD matrix is in-sample.}

\item{a}{Shape parameters for sigma^2. Either a positive real number or an
M-vector of positive real numbers.}

\item{b}{Shape parameter for 1-h^2. A positive real number.}

\item{coverage}{A number between 0 and 1 specifying the "coverage"
of the credible sets.}

\item{purity}{A number between 0 and 1 specifying the minimum 
absolute correlation allowed in a credible set.}
}
\value{
An h2D2 object. See \code{\link{h2D2-class}}.
}
\description{
Create an h2D2 object from GWAS summary statistics and an LD matrix estimate.
GWAS summary data, h2-D2 hyper-parameters, and useful variables are stored.
}
