% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Createh2D2Object.R
\name{Createh2D2Object}
\alias{Createh2D2Object}
\title{Create an h2D2 object.}
\usage{
h2D2 = Createh2D2Object(z,
                        R,
                        N,
                        SNP_ID = NULL,
                        trait = c("quantitative","binary"),
                        in_sample_LD = F,
                        a = 0.005,
                        b = 1e4,
                        coverage = 0.95,
                        purity = 0.5)
}
\arguments{
\item{z}{M-vector of z-scores.}

\item{R}{An M-by-M LD matrix that can be coerced to 
\code{\link[Matrix]{dsCMatrix-class}}.
The diagonal elements will be coerced to zeros.}

\item{N}{GWAS sample size. For binary traits, 'N' is the sum of the
number of cases and the number of controls.}

\item{SNP_ID}{Identifiers of SNPs. The default is c("SNP_1", ...).}

\item{trait}{Either "quantitative" or "binary".}

\item{in_sample_LD}{Whether the LD matrix is in-sample.}

\item{a}{Shape parameters for sigma^2. Either a positive real number or an
M-vector of positive real numbers.}

\item{b}{Shape parameter for 1-h^2. Must be a positive real number.
If the in-sample LD matrix is provided, we recommend setting b = NULL and
b will be estimated by a pre-training process before MCMC.}

\item{coverage}{A number between 0 and 1 specifying the "coverage"
of the credible sets.}

\item{purity}{A number between 0 and 1 specifying the minimum 
absolute correlation allowed in a credible set.}
}
\value{
An h2D2 object. See \code{\link{h2D2-class}}.
For quantitative traits, the input z-scores will be modified to
z / sqrt(1 + z^2/N - 1/N).
}
\description{
Create an h2D2 object from GWAS summary statistics and an LD matrix estimate.
GWAS z-scores, h2-D2 hyper-parameters, and useful variables are stored.
}
