% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2D2-class.R
\docType{class}
\name{h2D2-class}
\alias{h2D2-class}
\title{The h2D2 class}
\description{
The h2D2 object is used to store GWAS summary data, LD matrix, variables
useful for MCMC sampling, h2-D2 hyper-parameters, posterior samples, and
credible sets. Objects can be created by Createh2D2Object.
}
\section{Slots}{

\describe{
\item{\code{M}}{Number of SNPs.}

\item{\code{N}}{Number of individuals for quantitative traits.}

\item{\code{N1}}{Number of cases for binary traits.}

\item{\code{N0}}{Number of controls for binary traits.}

\item{\code{SNP_ID}}{IDs of SNPs. A character vector of size M.}

\item{\code{betaHat}}{An M-vector of marginal effect size estimates.}

\item{\code{S}}{An M-vector of standard errors of marginal effect size estimates.}

\item{\code{z}}{An M-vector of z-scores.}

\item{\code{R}}{An M-by-M correlation matrix in \code{\link[Matrix]{dsCMatrix-class}} 
format. The diagonal elements are set as 0.}

\item{\code{trait}}{A character string. One of "quantitative" or "binary".}

\item{\code{LD_pairs}}{A \code{\link[Matrix]{dgCMatrix-class}} object stored SNP 
pairs with |r|>=purity and transition probabilities used in MCMC algorithm to 
select SNP pair for switching.}

\item{\code{a}}{An M-vector of shape parameters for sigma^2.}

\item{\code{b}}{Shape parameter for 1-h^2.}

\item{\code{mcmc_samples}}{A list storing MCMC samples.}

\item{\code{mcmc_mean}}{A list storing posterior sample means.}

\item{\code{mcmc_sd}}{A list storing posterior sample standard errors.}

\item{\code{mcmc_quantile}}{A list storing posterior sample 2.5\%, 25\%, 50\%, 75\%, 
and 97.5\% quantiles.}

\item{\code{CL}}{An M-vector of credible levels of variables.}

\item{\code{coverage}}{A number between 0 and 1 specifying the "coverage" of the 
credible sets.}

\item{\code{purity}}{A number between 0 and 1 specifying the minimum absolute 
correlation allowed in a credible set.}

\item{\code{CS}}{A list of identified credible sets.}
}}

