% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2D2_MCMC.R
\name{h2D2_MCMC}
\alias{h2D2_MCMC}
\title{MCMC sampling under h2-D2 prior.}
\usage{
h2D2 = h2D2_MCMC(h2D2, mcmc_n = 5500, burn_in = 500, thin = 1, 
                 stepsize = 2, seed = 428, get_CS = TRUE, 
                 n_chain = 3, fold = 1.1,
                 pre_mcmc_n = 400, pre_burn_in = 200, pre_p = 0.05,
                 pre_maxiter = 10, pre_miniter = 2)
}
\arguments{
\item{h2D2}{An h2D2 object.}

\item{mcmc_n}{Number of posterior samples to generate.}

\item{burn_in}{Number of early samples to discard.}

\item{thin}{Thinning parameter of the MCMC chain. The default is 1, 
indicating no thinning.}

\item{stepsize}{Step size for proposal.}

\item{seed}{Random seed for MCMC sampling.}

\item{get_CS}{Logical. If get_CS = TRUE, detect CS after MCMC. Otherwise, 
CS will not be detected.}

\item{n_chain}{Number of parallel chains with different temperature.}

\item{fold}{Fold change of temperature between two neighbored parallel chains.
For instance, if \code{n_chain = 3} and \code{fold = 1.1}, the temperature
for the three parallel chains will be 1, 1.1, 1.21.}

\item{pre_mcmc_n}{Number of posterior samples to generate in each iteration
of pretraining. If b is given, the pretrain will not run.}

\item{pre_burn_in}{Number of early samples to discard in each iteration of
pretraining.}

\item{pre_p}{p-value threshold for the pretraining.}

\item{pre_maxiter}{Maximum number of pretraining iterations allowed.}

\item{pre_miniter}{Minimum number of pretraining iterations required.}
}
\value{
An h2D2 object with MCMC samples. See \code{\link{h2D2-class}}.
}
\description{
Using MCMC sampler to fit posterior distribution under h2-D2 prior.
}
