% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2D2_MCMC.R
\name{h2D2_MCMC}
\alias{h2D2_MCMC}
\title{MCMC sampling under h2-D2 prior.}
\usage{
h2D2_MCMC(h2D2, mcmc_n = 10000, burn_in = 5000, thin = 1, 
stepsize = 2, seed = 428)
}
\arguments{
\item{h2D2}{An h2D2 object.}

\item{mcmc_n}{Number of posterior samples to generate.}

\item{burn_in}{Number of early samples to discard.}

\item{thin}{Thinning parameter of the MCMC chain. The default is 1, 
indicating no thinning.}

\item{stepsize}{Step size for proposal.}

\item{seed}{Random seed for MCMC sampling.}
}
\value{
An h2D2 object with MCMC samples.
}
\description{
Using MCMC sampler to fit posterior distribution under h2-D2 prior.
}
